feat: refactor useScreenGuard hooks
